Переписан некрасивый код. Исправлен тестовый вывод.

+++ StringTime.h
@@ -4,6 +4,6 @@
 #include <string>
 
 /// @brief Отображение указанного времени в разговорной форме.
-std::wstring timeToString(int hour, int minutes, int seconds);
+std::wstring timeToString(uint8_t hour, uint8_t minutes, uint8_t seconds);
 
 #endif // STRINGTIME_H
